pyproject.toml
[project]
name = "fate-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Pavlukhin",email = "michael@pavlukhinlab.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "dspy (>=3.0.3,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1"
ruff = "^0.13.0"
mypy = "^1.18.1"
pre-commit = "^4.3.0"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

